jobScout/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "jobscout"
version = "0.0.1"
description = "Job discovery and fit scoring pipeline"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "selectolax>=0.3",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "typer>=0.12",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]